reorganizing the project according to the oficial documentation
