+++ app/templates/NewPlayer.html
