+++ app/templates/enter_game.html
